+++ stockinfo.py
@@ -1,75 +1,86 @@
 import streamlit as st
+import FinanceDataReader as fdr
 import pandas as pd
-import FinanceDataReader as fdr # 삼성전자 -> 종목코드를 알려주지 않음. 
+# import talib
+import streamlit.components.v1 as components
 import datetime
-import matplotlib.pyplot as plt
-import matplotlib 
-from io import BytesIO # 바이너리파일들을 읽고 쓸 때 사용하는 패키지
-import plotly.graph_objects as go
-import pandas as pd
+from fs import *
+# 대표 사이트 명
+st.title(':달러: 우리FISA 증권 :달러:')
+# Streamlit 제목 설정
+st.title(':안내데스크_직원::피부톤-2: 실시간 주식 종목 분석')
+# 사용자로부터 종목명, 종목코드 또는 티커 입력 받기
+ticker = st.text_input(':단안경을_쓴_얼굴: 종목코드 또는 종목 티커를 입력하세요:', 'AAPL')
+# TradingView 차트 삽입
+st.subheader(':막대_차트: Technical Overview')
+tradingview_widget = f"""
+<iframe src="https://www.tradingview.com/widgetembed/?symbol={ticker}&theme=dark&style=1&timezone=Asia/Seoul&withdateranges=1&hide_side_toolbar=1&allow_symbol_change=1&save_image=1&studies=[]&locale=kr" width="100%" height="600" frameborder="0" allowfullscreen></iframe>
+"""
+components.html(tradingview_widget, height=400)
+# FinanceDataReader를 사용하여 종목 데이터 가져오기
+data = fdr.DataReader(ticker, start='2024-01-01')
+# 실시간 주가 표시
+st.subheader('실시간 주가')
+st.write(f'현재가: {data.iloc[-1]["Close"]}')
+st.write(f'전날 종가: {data.iloc[-2]["Close"]}')
+st.write(f'최고가: {data["Close"].max()}')
+st.write(f'최저가: {data["Close"].min()}')
+# 과거 데이터 표시
+st.subheader('종목 히스토리')
+st.dataframe(data, width=1200)
+# 실시간 매수/매도 및 공매도 데이터를 가져오는 함수 (예시: 수동 계산)
+st.dataframe(bs1)
+st.dataframe(is1)
 
-# caching
-# 인자가 바뀌지 않는 함수 실행 결과를 저장 후 
-# 크롬의 임시 저장 폴더에 저장 후 재사용
-@st.cache_data
-def get_stock_info():
-    base_url =  "http://kind.krx.co.kr/corpgeneral/corpList.do"    
-    method = "download"
-    url = "{0}?method={1}".format(base_url, method)   
-    df = pd.read_html(url, header=0, encoding='cp949')[0] # MS cp949 / Windows 
-    df['종목코드']= df['종목코드'].apply(lambda x: f"{x:06d}")     
-    df = df[['회사명','종목코드']]
-    return df
-
-def get_ticker_symbol(company_name):     
-    df = get_stock_info()
-    code = df[df['회사명']==company_name]['종목코드'].values    
-    ticker_symbol = code[0]
-    return ticker_symbol
-
-with st.sidebar:
-    # stock_name을 입력받는 input창
-    stock_name = st.text_input('회사 이름을 입력하세요: ')
-
-
-    today = datetime.datetime.now()
-    this_year = today.year
-    jan_1 = datetime.date(this_year, 1, 1)
-
-    date_range = st.date_input(
-        "시작일과 종료일을 입력하세요",
-        (jan_1, today), # default 값
-        None,
-        today,
-        format="MM.DD.YYYY",
-    )
-
-    st.write(date_range)
-
-    accept = st.button('확인')
-
-if accept:
-    # 코드 조각 추가
-    ticker_symbol = get_ticker_symbol(stock_name)     
-    start_p = date_range[0]               
-    end_p = date_range[1] + datetime.timedelta(days=1) 
-    df = fdr.DataReader(f'KRX:{ticker_symbol}', start_p, end_p)
-    df.index = df.index.date
-    st.subheader(f"[{stock_name}] 주가 데이터")
-    st.dataframe(df.tail(7))
-
-    st.write(df)
-
-    fig = go.Figure(data=[go.Candlestick(x=df.index,
-                    open=df['Open'],
-                    high=df['High'],
-                    low=df['Low'],
-                    close=df['Close'])])
-
-    st.plotly_chart(fig, use_container_width=True)
-
-    excel_data = BytesIO()      
-    df.to_excel(excel_data)
-
-    st.download_button("엑셀 파일 다운로드", 
-            excel_data, file_name='stock_data.xlsx')
+def get_trade_data(ticker):
+    # 예시로 FinanceDataReader를 사용하여 데이터를 가져옵니다.
+    # 실제로는 거래소 API 등을 사용하여 매수/매도 데이터를 가져옵니다.
+    data = fdr.DataReader(ticker, start='2024-01-01')
+    return data
+# 최근 한달간 매수/매도 및 공매도 데이터 가져오기
+end_date = datetime.date.today()
+start_date = end_date - datetime.timedelta(days=30)
+trade_data = get_trade_data(ticker)
+# 여기서는 주식 데이터만으로 매수/매도량을 계산할 수 있는 예시입니다.
+# 실제 매수/매도량은 거래소 API 등을 사용하여 가져와야 합니다.
+institution_buy = trade_data['Volume'].sum()  # 예시 데이터: 기관의 최근 한달간 총 매수량
+institution_sell = trade_data['Volume'].sum() * 0.5  # 예시 데이터: 기관의 최근 한달간 총 매도량
+individual_buy = trade_data['Volume'].sum() * 0.5  # 예시 데이터: 개인투자자의 최근 한달간 총 매수량
+individual_sell = trade_data['Volume'].sum() * 0.4  # 예시 데이터: 개인투자자의 최근 한달간 총 매도량
+short_selling = trade_data['Volume'].sum() * 0.1  # 예시 데이터: 공매도 현황
+# 종합 분석
+if institution_buy > institution_sell and individual_buy > individual_sell:
+    opinion = '매수 의견'
+    opinion_description = '기관과 개인 투자자 모두 최근 한달간 매수량이 매도량을 초과하므로, 해당 종목의 주식 가격 상승 가능성이 높다고 판단됩니다.'
+elif institution_sell > institution_buy and individual_sell > individual_buy:
+    opinion = '매도 의견'
+    opinion_description = '기관과 개인 투자자 모두 최근 한달간 매도량이 매수량을 초과하므로, 해당 종목의 주식 가격 하락 가능성이 높다고 판단됩니다.'
+else:
+    opinion = '중립 의견'
+    opinion_description = '기관과 개인 투자자의 매수량과 매도량이 비슷하므로, 해당 종목의 주식 가격이 변동 없이 안정적인 상태일 가능성이 높습니다.'
+# 종합 분석 결과 표시
+st.subheader('종합 분석 결과')
+st.write(f'기관의 최근 한달간 총 매수량: {institution_buy}')
+st.write(f'기관의 최근 한달간 총 매도량: {institution_sell}')
+st.write(f'개인투자자의 최근 한달간 총 매수량: {individual_buy}')
+st.write(f'개인투자자의 최근 한달간 총 매도량: {individual_sell}')
+st.write(f'공매도 현황: {short_selling}')
+st.write(f'현재 주식 가격에 대한 의견: {opinion}')
+st.write(f'의견 설명: {opinion_description}')
+# 인기 종목과 주가 변동성을 표로 표시
+st.subheader('실시간 인기 종목 및 주가 변동성')
+# 실시간 인기 종목 데이터 가져오기
+krx = fdr.StockListing('KRX')
+popular_stocks = krx.head(5)  # 상위 5개 종목 예시
+# 인기 종목 실시간 데이터 가져오기 및 변동성 계산
+popular_stocks_data = []
+for code in popular_stocks['Code']:  # 'Symbol' 대신 'Code' 사용
+    stock_data = fdr.DataReader(code, start='2023-01-01')
+    current_price = stock_data.iloc[-1]['Close']
+    prev_close = stock_data.iloc[-2]['Close']
+    volatility = ((current_price - prev_close) / prev_close) * 100
+    popular_stocks_data.append([code, current_price, prev_close, volatility])
+# 데이터프레임 생성
+popular_stocks_df = pd.DataFrame(popular_stocks_data, columns=['티커', '현재가', '전날종가', '변동성'])
+# 표로 표시
+st.table(popular_stocks_df)
